// Frontend/src/components/UI/Minimal/types/user.ts
export type ICurriculumItem = {
  id?: number;
  instrument_id?: number;
  months: number;
  weekly_classes: number;
  total_classes: number;
  name: string;
  price: number;
  instrument_price: number;
  instrument_rental_fee: number;
  instrument_discount: number;
  image_file?: string;
  additional_text?: string;
  is_deleted: string;
  created_date: Date;
	instrument?:IInstrumentItem;
}





// 악기
export type IInstrumentItem = {
  id?: number;
  name: string;
  image_file1?: string;
  image_file2?: string;
  display_order: number;
  is_deleted: string;
  created_date: Date;
}